+++ NEWS.md
@@ -3,6 +3,12 @@ hello
 
 JSON-RPC API toolkit
 
+Version 3.0.0 - xx Mar 2013
+---------------------------
+
+* Merge hello2 in hello
+* Fix bug by supervising of cowboy listeners
+
 Version 2.1.1 - xx Nov 2013
 ---------------------------